Fix rendering of docstrings in Sphinx
For now we use RST, as it is simpler (it doesn't require code fences on the
examples), and with default_role = 'code', it is effectively the same as
Markdown. But I have also included some code that will let us switch to
Markdown in the future if we want to.