Refactor Tuple.reduce()
The existing Tuple.reduce() has been renamed to Tuple.expand() (pending a better
name).

Tuple.reduce() simplifies the Tuple as much as possible, for example, by
merging entries into ellipses. Tuple.expand(), on the other hand, expands a
Tuple so that there are no ellipses in the result and the .args always has as
many elements as the shape.