Remove the positive_slices hypothesis strategy
We no longer assume(False) on NotImplementedError, so this is not needed to
avoid health checks. Rather, it is better to test all slices so that we can be
sure that cases that should give NotImplementedError do so correctly, and so
that we don't accidentally miss some cases that actually are supported.