Add shape parameter to isempty() and move the logic to the subclasses
The shape parameter is needed because reduction isn't enough to guarantee a
correct answer, since the array itself could be empty due to a 0 in a
non-first dimension. For example, Slice(0, 10) will give an empty result for a
shape (10, 0) array even though Slice(0, 10).reduce((10, 0)) returns Slice(0,
10). The issue is that len(Slice) gives the length of the axis being sliced,
not the total array, which is what isempty() is designed to do.