Make the README a separate file from index.md
If they are the same file, any relative links in the index.md will cause a 404
in the README on GitHub. So instead we make the README a much shorter page
with some simple examples, and keep the index.md as it is for the
documentation.

This also updates the doctester to properly doctest the separate README.