Make as_subindex raise ValueError in some cases when the indices do not intersect
Previously we gave an empty slice, but this would result in arrays whose
shapes do not match, or worse, in the case where one index produced a scalar,
an index that produces an IndexError. Since the "correct" index to give in
this case is either impossible to represent or cannot be represented with the
indices that are currently implemented (see the comments in the code), we
instead raise ValueError.

The ValueError is not raised in every instance where the indices do not
intersect. For example Slice.as_subindex(Slice) never raises ValueError. It is
only in the cases where an Integer index is involved, where the number of
dimensions is reduced, which makes it impossible to return a correct index.

Also add a note about symmetry to the as_subindex docstring.