Properly handle IntegerArrays that do not do bounds checks
NumPy does not do bounds checks on integer arrays if one of the non-indexed
axes is 0. This behavior will be deprecated in NumPy 1.20, at which point we
will change the behavior to raise an IndexError. But until 1.20 is released,
we cannot easily test against the post-deprecation behavior, as there are no
warning or errors from NumPy to check against, so for now, we match the 1.19
behavior.