Use SymPy to compute ceilings
int(math.ceil(a/b)) is incorrect if a/b is too large, because precision is lost in
the least significant digit. Instead use sympy.Rational.ceiling(), which is
exact.

This doesn't *really* matter, because it only affects indices that are out of
bounds for any possible array, but hypothesis generates tests with large
integers, which causes failures otherwise.