Handle a NumPy bug in the handling of some boolean arrays
NumPy has a fast path that incorrectly compares the sizes of the array and
boolean index, rather than the shapes. As a result, invalid indices can be
accepted in some cases when the size but not the shapes match, and in other
cases it gives the wrong exception. This is being fixed in the upstream NumPy
(see the issue numbers linked in the comments). Rather than matching the
broken NumPy behavior here, we do the right thing, and guard against the
broken NumPy behavior in the tests.