Raise ValueError for tuple arguments to Tuple instead of NotImplementedError
These only work in NumPy as fancy indices. However, there is too much of a
risk of API confusion in thinking that Tuple() works like Tuple((1, 2, 3))
instead of Tuple(1, 2, 3). So we make this an unconditional error with helpful
error messages. Fancy indices should use either a list or array.

c.f. issue #17.