Disallow booleans in operator.index() contexts
This adds a new helper function operator_index() that works like
operator.index() except it disallows booleans. The built-in bool works with
operator.index, but np.bool_ gives a deprecation warning, so according to our
policy of making errors out of things that give deprecation warnings in NumPy,
we make this give an error. This affects Integer(), Slice(), and asshape().

A slight compatibility break here against NumPy is that NumPy actually does
allow booleans in slices. However, I'm not too worried about this as it's poor
form to do that, and most likely would indicate a bug in user code.