Refactor check_same() in the tests to more naturally test functions that don't return an index
Not only is this cleaner, as we don't have to "workaround" the function to
pretend that it is testing an index, it fixes an issue where the previous way
was not actually testing if a function raised an exception properly, because
it would just be indexed anyway after calling the function. This affected
newshape() and isempty() tests.