Allow Tuples of IntegerArrays and other index types
Any slices, ellipses, or newaxes must not be between the arrays, or it raises
NotImplementedError. This will probably never be implemented, because it's a
bit of a wart case in the indexing API, which should raise an error (or at
least that's what Travis told me to do).

Also makes Tuple.expand() convert Integers into IntegerArrays as arrays when
the tuple also contains arrays (so that they will also be broadcast).