Fix Tuple.newshape() to properly handle array indices
This adds a private _axis keyword argument to newshape() methods. This is
needed to keep the NumPy 1.19 behavior of not raising IndexError on out of
bounds indices on empty arrays in some cases. This is needed because the
specific cases when this happens require knowing the entire shape, but
previously we just passed through idx.newshape(shape[i]) in Tuple.newshape().
This behavior will be deprecated in NumPy 1.20, at which point we will stop
supporting it (breaking support earlier is hard because we cannot explicitly
test against NumPy without a deprecation warning to catch). When this happens,
the _axis argument to the newshape() methods will be removed and we can go
back to just passing through shape[i] as before.