Fix lots of issues with as_subindex on array indices
The previous logic didn't properly handle masking multiple arrays
simultaneously.

The currently logic may still have some bugs (e.g., I'm not sure if it's not
possible for it to try to create a Tuple that raises an IndexError because it
can't broadcast, because the integer array results are always flat).