Make as_subindex return non-boolean array indices when possible.
This fixes some bugs where the shapes of idx1.as_subindex(idx2) and
idx2.as_subindex(idx1) were incompatible. It is nice to make the indices not
boolean when possible so that the resulting axes are not flattened, but it
must be done with care so that the two as_subindex directions match up in
shape. I'm not fully confident yet that this is all working correctly, but I
know at least that certain cases that weren't working before are now working
correctly.