Make hash(idx.raw) == hash(idx) when idx.raw is hashable
That way it is compatible with idx == idx.raw.

Note that idx == x but hash(idx) != hash(x) is still possible for instances
where idx.raw is nonhashable. For example, the raw form for True and False is
array(True) and array(False), which are not hashable. We could try to detect
these cases, but for now, we don't.

Fixes #88.