Make as_subchunks() only return the chunk indices
The as_subindex() indices are no longer computed as part of the calculation,
so returning them is an extra computation that may not actually be needed.
Furthermore, they can easily be calculated directly with idx.as_subindex(c) by
the caller.

Also improve the docstrings of as_subchunks() and as_subindex() with respect
to this change.