Define __eq__ on each individual class
This speeds up == comparisons considerably, as each class can do an exact
comparison against the classes it knows it can be equal to, rather than
calling ndindex() on any unknown object. The generic NDIndex.__eq__ method is
left intact as a fallback, although currently it is unused by any code.

This also adds __hash__ to each class, as Python requires any class that
redefines __eq__ to also redefine __hash__ in order to be hashable, even if
__hash__ is already defined on the superclass.