Test that the fallback isn't required in as_subchunks()
It should only ever be called in cases where it would raise
NotImplementedError. The private _force_slow flag now defaults to None, which
does the fallback when the fast case isn't implemented. If it is False, it
raises a RuntimeError if the fallback is called. This allows us to test that
this never happens, except in the cases where it does give
NotImplementedError.

Once all as_subindex() and as_subchunks() cases are completely implemented (no
more NotImplementedErrors), the fallback should never be called, except for
manually by _force_slow=True in the tests.