Move imports from inside functions to the bottom of the file
They can't be at the top because it creates circular import issues. But having
them inside the functions can be a performance hit that is significant if the
function is called many times. Since having them at the bottom of the file
fixes the circular import problem just as well, this is preferable.