Remove codecov
We require 100% coverage and print the report in the Travis log anyway, which
generally only fails with a single line or two. So codecov isn't very useful,
and I haven't used it at all since adding it. Given that it had a recent
security incident https://about.codecov.io/security-update/, I think we should
remove it entirely.