Use slice(None) in iter_indices()
This makes it easier to reuse the index for similarly shaped arrays. For
example, when using iter_indices(shape, (-2, -1)) to iterate stacks of
matrices, the same indices can be use to index the transpose of those
matrices. A specific slice can still be obtained by using expand() or reduce()
on the index with the array shape.