Remove SymPy as a dependency
Instead we just copy the relevant Chinese Remainder Theorem code into ndindex.
This means a minor performance loss given that SymPy's crt() uses gmpy2. But
this is worth it because SymPy is a pretty heavy dependency for such a simple
function, which itself is only used in as_subindex() for slices with step size
greater than 1.