Be more robust with testing against deprecated numpy functionality
Previously we only relied on filterwarnings=error in pytest.ini to turn
deprecation warnings into errors. We now also do this explicitly in the
check_same() function. Also add pytest.ini and conftest.py to the MANIFEST.in
so that they are included in the sdist. We still use filterwarnings=error in
pytest.ini because we want to make sure we aren't using any other deprecated
behavior anywhere else.

Fixes #150.