Disallow skip_axes that mix positive and negative integers (for now)
Given the updated logic where skip_axes apply before broadcasting, this is too
difficult to handle correctly, due to the inherent ambiguity where a positive
and negative index could refer to the same value. It's especially difficult to
do test generation in this case, and I can't figure out how to do it. Given
that this isn't really something that I need support for, I'm going to leave
it unimplemented for now.