Update mutually_broadcastable_shapes_with_skip_axes to generate lists of skip_axes
This changes the way the strategy is generated to construct skip_axes and a
result shape first, then generate shapes that match those skip axes and
broadcast to the result shape. It doesn't attempt to directly use
mutually_broadcastable_shapes. This is unfortunate, and it does require a
little bit of filtering, but it is the only way that I could figure out how to
generate the full range of possible examples. skip_axes_st is now always
generated in conjunction with this strategy, as 1) generating the skip axes
first and then the shapes doesn't seem to work, and 2) the two strategies are
always used together anyway.

This doesn't yet update all the tests that use this (nor are the functions
themselves updated), but test_mutually_broadcastable_shapes_with_skip_axes()
passes.