No longer add None indices in the return value of broadcast_shapes with skip_axes
This is impossible to do in general if we allow both negative and nonnegative
skip_axes, and it's also too complex to check the result shape skip axes to
make sure they are consistent.

Unfortunately, this is a backwards compatibility break, but hopefully no one
is actually using this function yet.