Update places that use ndindex() to use ndindex[]
I've primarily only done this in places where the index is a literal
(particularly a slice or tuple literal, which is the only place where it
matters). In other places, especially in the code, a generic ndindex(idx) is
clearer as a constructor than ndindex[idx], so I've left it alone.